refactor(dbns): split dbns into core / manipulators / iostream / debug (#1334)
Fourth type of Phase 2 group 5a (#1444), and the first in the group where the
core's I/O was structural rather than an include.

    core.hpp        73,581 lines / 361 headers / 0 I/O-family
    dbns.hpp        82,967 lines / 392 headers / 5 I/O-family  (was 83,961 / 406 / 5)

Zero I/O-family headers is met. Under-45,000 lines is not, and reporting the
number rather than moving the bar: internal/blocktriple/blocktriple.hpp, which
dbns converts through, is 69,758 lines on its own (already 0 I/O), so this core
is blocktriple plus about 3,800 lines. The ceiling belongs to the substrate.

Four things had to move, and only the first is an include shuffle:

1. to_binary() formats through a stringstream -> manipulators.hpp.

2. operator<< and operator>> were in-class friend DEFINITIONS. An in-class friend
   definition pins <iostream> into every translation unit that includes the
   class -- the trap blockdigit and floatcascade were fixed for. Both go through
   dbns's public double conversion and assignment and need no private access, so
   they are now ordinary namespace-scope templates in iostream.hpp rather than
   friends in the class.

3. debugConstexprParameters() -- 22 std::cout statements in a member function --
   is declared in the class and defined out-of-line in debug.hpp. Same shape
   blocktriple's constexprClassParameters() got in #1388: forgetting the include
   is a link error naming the missing function, which is the better failure.

4. convert_ieee754's (a,b) search traced behind `if constexpr (bDebug)`. A
   discarded if-constexpr branch still requires std::cout to be DECLARED, so the
   three trace sites are now behind a DBNS_TRACE_CONVERSION preprocessor switch.

   bDebug was a local `constexpr bool bDebug = false;` that no caller could
   reach -- tracing could not be enabled from outside at all, the same defect
   #1387 found in the block layer. It is now switchable, and switched on it
   produces byte-identical text (see below).

Also: operator<<(ostream&, DbnsArithmeticStatistics) is declared in the core and
defined in iostream.hpp; native/ieee754.hpp -> native/ieee754_core.hpp;
DBNS_ENABLE_LITERALS, DBNS_THROW_ARITHMETIC_EXCEPTION and the
QUIRE_THROW_ARITHMETIC_EXCEPTION cascade default in dbns_impl.hpp (#1436);
manipulators.hpp and attributes.hpp now name the headers they use (#1389).

manipulators.hpp does NOT include iostream.hpp: unlike posit's, dbns's
pretty_print/info_print build from the bit pattern rather than streaming the
value, so the dependency runs one way only.

Noted, deliberately unchanged: debugConstexprParameters() reports the masks with
to_binary(mask, bitsInBlock), and that overload's second parameter is
`bool bNibbleMarker`, not a width -- so bitsInBlock has always been read as
`true` and the width has always defaulted to 0. Correcting it would change
output, so it stays as it is.

Verification:

- Layer gate: core.hpp alone, and core.hpp + manipulators.hpp, both compile,
  link and run under -Wall -Wpedantic on gcc 13.3 and clang.
- Self-containment: core, manipulators, iostream, debug, attributes and the
  umbrella each compile standalone with 0 errors on both compilers.
- 16 targets built and run green on both compilers, including dbns_table, which
  calls debugConstexprParameters() and so proves the out-of-line definition
  links, and quire_dbns_fdp for the QUIRE_THROW cascade.
- Differential against main: five dbns configurations swept over their entire
  encoding space (dbns<8,3>, <10,4>, <12,5>, <9,4,uint32> exhaustively, <16,8>
  over 65,536), each rendered by to_binary, to_hex and pretty_print in every
  flag combination, info_print, color_print, operator<< under four stream
  states, the native conversions, plus operator>> over 10 texts, a dynamic-range
  round trip, the relocated debugConstexprParameters() report and the
  DbnsArithmeticStatistics report. 29,992,863 bytes byte-identical -- old vs new
  on gcc, old vs new on clang, and gcc vs clang.
- Trace text: main cannot switch tracing on, so the comparison was made by
  setting bDebug = true in a worktree of main and running it against
  -DDBNS_TRACE_CONVERSION=1 here. 1,273,937 bytes over 125 conversions
  byte-identical on both compilers.
